+++ src/Ext/Side/Body.h
@@ -44,6 +44,7 @@ public:
 		PhobosPCXFile SuperWeaponSidebar_TopPCX;
 		PhobosPCXFile SuperWeaponSidebar_CenterPCX;
 		PhobosPCXFile SuperWeaponSidebar_BottomPCX;
+		Nullable<char*> EVA_Tag;
 
 		ExtData(SideClass* OwnerObject) : Extension<SideClass>(OwnerObject)
 			, ArrayIndex { -1 }
@@ -74,6 +75,7 @@ public:
 			, SuperWeaponSidebar_TopPCX {}
 			, SuperWeaponSidebar_CenterPCX {}
 			, SuperWeaponSidebar_BottomPCX {}
+			, EVA_Tag { }
 		{ }
 
 		virtual ~ExtData() = default;
@@ -98,6 +100,9 @@ public:
 	};
 
 	static ExtContainer ExtMap;
+
 	static bool LoadGlobals(PhobosStreamReader& Stm);
 	static bool SaveGlobals(PhobosStreamWriter& Stm);
+
+	static void UpdateMainEvaVoice(BuildingClass* pThis);
 };